requirements.txt
astropy==4.0.1.post1
influxdb==5.2.3
matplotlib==3.1.1
numpy==1.16.3
Pillow==6.0.0
pyserial==3.4
scikit-image==0.15.0
scipy==1.3.1
simple-pid==0.2.3
pyindi-client==0.2.5  # also requires libindi-dev (apt)
